+++ changelog.md
@@ -9,6 +9,9 @@
 -   Feature: Add redux actions to change language and add it on copyright symbol and copyright year on footer.
 -   Bugfix: Add default value for onSubmit on SignupForm.
 -   Feature: Add terms checkbox on SignupForm.
+-   Feature: Add default redirect routes in routes config.
+-   Feature: Update auth roles and use them on user slice and internal pages.
+-   Feature: Push to correct default page depending on role after login/register on landing page.
 
 ### v0.6.2